# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Taco
  description: 'Complete documentation for Taco API server.'
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000'
paths:
...
  /api/notice/list:
    get:
      summary: ListAll.
      operationId: listAll
      description: "Récupère la liste des avis de l'utilisateur qui émet la requette\nselon une page et un nombre qu'il peut définir."
      parameters:
        -
          in: query
          name: page
          description: 'numéro correspondant à la page à lire.'
          example: 1
          required: false
          schema:
            type: integer
            description: 'numéro correspondant à la page à lire.'
            example: 1
        -
          in: query
          name: quantity
          description: 'nombre d''éléments à lire par page.'
          example: 10
          required: false
          schema:
            type: integer
            description: 'nombre d''éléments à lire par page.'
            example: 10
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer ${token}'
          schema:
            type: string
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: 'Wrong number of segments'
                properties:
                  error:
                    type: string
                    example: 'Wrong number of segments'
      tags:
        - 'Route pour les avis'
...
  /api/trip/list:
    get:
      summary: ListAll.
      operationId: listAll
      description: "Récupère la liste des déplacement fait par l'utilisateur qui émet la requette\nselon une page et un nombre qu'il peut définir."
      parameters:
        -
          in: query
          name: page
          description: 'numéro correspondant à la page à lire.'
          example: 1
          required: false
          schema:
            type: integer
            description: 'numéro correspondant à la page à lire.'
            example: 1
        -
          in: query
          name: quantity
          description: 'nombre d''éléments à lire par page.'
          example: 10
          required: false
          schema:
            type: integer
            description: 'nombre d''éléments à lire par page.'
            example: 10
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer ${token}'
          schema:
            type: string
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    -
                      id: 1
                      to: Maison
                      to_lat: 3.85
                      to_lon: 11.37
                      from: 'Poste Centrale'
                      from_lat: 3.9
                      from_lon: 11.46
                      distance: 45
                      price: 5000
                      status: broadcasting
                      type: depot
                      created_at: null
                      client:
                        id: 11
                        first_name: 'Adah Brakus'
                        last_name: 'Filomena Tromp'
                      driver:
                        id: null
                        first_name: null
                        last_name: null
                    -
                      id: 1
                      to: Maison
                      to_lat: 3.85
                      to_lon: 11.37
                      from: 'Poste Centrale'
                      from_lat: 3.9
                      from_lon: 11.46
                      distance: 45
                      price: 5000
                      status: broadcasting
                      type: depot
                      created_at: null
                      client:
                        id: 11
                        first_name: 'Adah Brakus'
                        last_name: 'Filomena Tromp'
                      driver:
                        id: null
                        first_name: null
                        last_name: null
                properties:
                  data:
                    type: array
                    example:
                      -
                        id: 1
                        to: Maison
                        to_lat: 3.85
                        to_lon: 11.37
                        from: 'Poste Centrale'
                        from_lat: 3.9
                        from_lon: 11.46
                        distance: 45
                        price: 5000
                        status: broadcasting
                        type: depot
                        created_at: null
                        client:
                          id: 11
                          first_name: 'Adah Brakus'
                          last_name: 'Filomena Tromp'
                        driver:
                          id: null
                          first_name: null
                          last_name: null
                      -
                        id: 1
                        to: Maison
                        to_lat: 3.85
                        to_lon: 11.37
                        from: 'Poste Centrale'
                        from_lat: 3.9
                        from_lon: 11.46
                        distance: 45
                        price: 5000
                        status: broadcasting
                        type: depot
                        created_at: null
                        client:
                          id: 11
                          first_name: 'Adah Brakus'
                          last_name: 'Filomena Tromp'
                        driver:
                          id: null
                          first_name: null
                          last_name: null
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 1
                        to:
                          type: string
                          example: Maison
                        to_lat:
                          type: number
                          example: 3.85
                        to_lon:
                          type: number
                          example: 11.37
                        from:
                          type: string
                          example: 'Poste Centrale'
                        from_lat:
                          type: number
                          example: 3.9
                        from_lon:
                          type: number
                          example: 11.46
                        distance:
                          type: integer
                          example: 45
                        price:
                          type: integer
                          example: 5000
                        status:
                          type: string
                          example: broadcasting
                        type:
                          type: string
                          example: depot
                        created_at:
                          type: string
                          example: null
                        client:
                          type: object
                          properties:
                            id:
                              type: integer
                              example: 11
                            first_name:
                              type: string
                              example: 'Adah Brakus'
                            last_name:
                              type: string
                              example: 'Filomena Tromp'
                        driver:
                          type: object
                          properties:
                            id:
                              type: string
                              example: null
                            first_name:
                              type: string
                              example: null
                            last_name:
                              type: string
                              example: null
      tags:
        - 'Routes pour les déplacements'
...
  /api/users/nearby:
    get:
      summary: ListAll.
      operationId: listAll
      description: 'Les utilisateurs retournés sont fonction du périmètre passé en paramètre.'
      parameters:
        -
          in: query
          name: distance
          description: 'distance autour de l''utilisateur qui fait la demande. Default is 2000m/2km.'
          example: 1500
          required: false
          schema:
            type: integer
            description: 'distance autour de l''utilisateur qui fait la demande. Default is 2000m/2km.'
            example: 1500
        -
          in: query
          name: lat
          description: 'latitude de l''utilisateur qui fait la demande.'
          example: 3.851248
          required: false
          schema:
            type: number
            description: 'latitude de l''utilisateur qui fait la demande.'
            example: 3.851248
        -
          in: query
          name: lon
          description: 'longitude de l''utilisateur qui fait la demande.'
          example: 11.154879
          required: false
          schema:
            type: number
            description: 'longitude de l''utilisateur qui fait la demande.'
            example: 11.154879
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer ${token}'
          schema:
            type: string
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    -
                      distance: null
                      email: neal.bartoletti@example.com
                      first_name: 'Fae Nienow'
                      id: 57
                      latitude: 3.485089
                      longitude: 11.50925
                      last_name: 'Alena Romaguera V'
                      phone: '+18433270167'
                      points: 56472
                      picture: true
                      rule: []
                      city:
                        id: 2
                        country: Cameroon
                        city: Douala
                        country_code: CM
                    -
                      distance: null
                      email: hill.stuart@example.net
                      first_name: 'Guadalupe Greenfelder'
                      id: 58
                      latitude: 3.120746
                      longitude: 11.998619
                      last_name: 'Kiarra Walker'
                      phone: '+12402917764'
                      points: 12325
                      picture: true
                      rule: []
                      city:
                        id: 4
                        country: Sudan
                        city: 'New Jacquelynport'
                        country_code: YE
                properties:
                  data:
                    type: array
                    example:
                      -
                        distance: null
                        email: neal.bartoletti@example.com
                        first_name: 'Fae Nienow'
                        id: 57
                        latitude: 3.485089
                        longitude: 11.50925
                        last_name: 'Alena Romaguera V'
                        phone: '+18433270167'
                        points: 56472
                        picture: true
                        rule: []
                        city:
                          id: 2
                          country: Cameroon
                          city: Douala
                          country_code: CM
                      -
                        distance: null
                        email: hill.stuart@example.net
                        first_name: 'Guadalupe Greenfelder'
                        id: 58
                        latitude: 3.120746
                        longitude: 11.998619
                        last_name: 'Kiarra Walker'
                        phone: '+12402917764'
                        points: 12325
                        picture: true
                        rule: []
                        city:
                          id: 4
                          country: Sudan
                          city: 'New Jacquelynport'
                          country_code: YE
                    items:
                      type: object
                      properties:
                        distance:
                          type: string
                          example: null
                        email:
                          type: string
                          example: neal.bartoletti@example.com
                        first_name:
                          type: string
                          example: 'Fae Nienow'
                        id:
                          type: integer
                          example: 57
                        latitude:
                          type: number
                          example: 3.485089
                        longitude:
                          type: number
                          example: 11.50925
                        last_name:
                          type: string
                          example: 'Alena Romaguera V'
                        phone:
                          type: string
                          example: '+18433270167'
                        points:
                          type: integer
                          example: 56472
                        picture:
                          type: boolean
                          example: true
                        rule:
                          type: array
                          example: []
                        city:
                          type: object
                          properties:
                            id:
                              type: integer
                              example: 2
                            country:
                              type: string
                              example: Cameroon
                            city:
                              type: string
                              example: Douala
                            country_code:
                              type: string
                              example: CM
      tags:
        - 'Routes pour utilisateurs'
...